requirements.txt
streamlit==1.32.0
groq==0.4.2
supabase==2.3.7
python-dotenv==1.0.1
google-genai==0.3.0
postgrest==0.10.7
httpx==0.27.0
